terraform: enable conditions to evaluate warnings
As per b133cfde, check rules now support a `warning_message`, which can also be applied to input variable validation. This commit introduces the ability to raise a warning diagnostic when a check rule is configured to issue a warning. Since the function caller determines the severity in certain contexts, this implementation includes logic to downgrade error severities to warnings when explicitly overridden by the user.